fix: speed up initial client bundling
