+++ client-src/utils/log.js
@@ -18,23 +18,4 @@ setLogLevel(defaultLevel);
 
 const log = logger.getLogger(name);
 
-const logEnabledFeatures = (features) => {
-  const enabledFeatures = Object.keys(features);
-  if (!features || enabledFeatures.length === 0) {
-    return;
-  }
-
-  let logString = "Server started:";
-
-  // Server started: Hot Module Replacement enabled, Live Reloading enabled, Overlay disabled.
-  for (let i = 0; i < enabledFeatures.length; i++) {
-    const key = enabledFeatures[i];
-    logString += ` ${key} ${features[key] ? "enabled" : "disabled"},`;
-  }
-  // replace last comma with a period
-  logString = logString.slice(0, -1).concat(".");
-
-  log.info(logString);
-};
-
-export { log, logEnabledFeatures, setLogLevel };
+export { log, setLogLevel };
