--- client-src/utils/parseURL.js
@@ -1,70 +1,0 @@
-/**
- * @returns {string}
- */
-function getCurrentScriptSource() {
-  // `document.currentScript` is the most accurate way to find the current script,
-  // but is not supported in all browsers.
-  if (document.currentScript) {
-    return document.currentScript.getAttribute("src");
-  }
-
-  // Fallback to getting all scripts running in the document.
-  const scriptElements = document.scripts || [];
-  const scriptElementsWithSrc = Array.prototype.filter.call(
-    scriptElements,
-    (element) => element.getAttribute("src"),
-  );
-
-  if (scriptElementsWithSrc.length > 0) {
-    const currentScript =
-      scriptElementsWithSrc[scriptElementsWithSrc.length - 1];
-
-    return currentScript.getAttribute("src");
-  }
-
-  // Fail as there was no script to use.
-  throw new Error("[webpack-dev-server] Failed to get current script source.");
-}
-
-/**
- * @param {string} resourceQuery
- * @returns {{ [key: string]: string | boolean }}
- */
-function parseURL(resourceQuery) {
-  /** @type {{ [key: string]: string }} */
-  let options = {};
-
-  if (typeof resourceQuery === "string" && resourceQuery !== "") {
-    const searchParams = resourceQuery.slice(1).split("&");
-
-    for (let i = 0; i < searchParams.length; i++) {
-      const pair = searchParams[i].split("=");
-
-      options[pair[0]] = decodeURIComponent(pair[1]);
-    }
-  } else {
-    // Else, get the url from the <script> this file was called with.
-    const scriptSource = getCurrentScriptSource();
-
-    let scriptSourceURL;
-
-    try {
-      // The placeholder `baseURL` with `window.location.href`,
-      // is to allow parsing of path-relative or protocol-relative URLs,
-      // and will have no effect if `scriptSource` is a fully valid URL.
-      scriptSourceURL = new URL(scriptSource, self.location.href);
-    } catch (error) {
-      // URL parsing failed, do nothing.
-      // We will still proceed to see if we can recover using `resourceQuery`
-    }
-
-    if (scriptSourceURL) {
-      options = scriptSourceURL;
-      options.fromCurrentScript = true;
-    }
-  }
-
-  return options;
-}
-
-export default parseURL;
